Conflicts:
	src/include/AmazonSDKUtil.h

+++ src/include/S3/S3MultipartUploadOperation_Internal.h
@@ -48,6 +48,6 @@ typedef void (^AbortMultipartUploadBlock)();
 @property (nonatomic, retain) AmazonServiceResponse *response;
 @property (nonatomic, retain) NSError *error;
 @property (nonatomic, retain) NSException *exception;
-@property (nonatomic, assign) NSUInteger partSize;
+@property (nonatomic, assign) int64_t partSize;
 
 @end
